Finalized Game
Included in my assets folder is the game executable. Some issues that I
am too tired to fix, and are honestly not that big of a deal, are that
the player health shows up at the credits scene, and both the background
for the camera and end scene are still blue.