Added an extra idle frame to the beginning of all player attack animations
Extra frame needed because Unity loops the attack animation back to the very first frame, which originally was when the player started to swing the sword. This effectively made the player attack twice as the initial swing has the collider which enemies can get hit by. The idle frame prevents this.