Added freezeAnimationMovement to EnemyChasePlayer script
Added this function in order for goblin to freeze for a second before being able to attack again, as using this function in EnemyMovement did not work as the script was disabled while goblin was chasing player. So when the goblin stopped attacking it faced a different direction immediately after as it was using the last direction it faced in EnemyMovement, rather than the last direction it faced when chasing the player.