Removed call to freezeAnimationMovement in turnOffEnemyAggro function in EnemyChasePlayer script
This call to freezeAnimationMovement as well as the same function call in GoblinAttack script seemed to cause the goblin to freeze for too long, even after the attack cooldown was done. As such, the goblin would move because the attack cooldown timer was done, but its animation would remain idle.