Fixed issue where dialogue box would not display proper dialogue
Issue was that the dialogue manager did not initialize the objectNameText and dialogueText to have values in the script, so the values had to be dragged in from the Unity editor itself. Changed it so now the script looks for the text gameObject instead of having to drag it in from the editor.