Created DontDestroyOnLoadManager static class to handle all gameObjects that have DontDestroyOnLoad, and to properly destroy those gameObjects.
Before, had a script called DestroyAll to destroy gameObjects that had DontDestroyOnLoad. Problem arose when game was built because method of destroying gameObjects was to create a temporary gameObject, calling DontDestroyOnLoad on it, the iterating through all root objects using that temporary gameObject to destroy them. This works in the editor, but not in final build. Another issue was that despite being destroyed, if gameObjects' static 'Exist' variable was not set to 'false' before destroying, Unity would think it still exists even after being destroyed. The static class DontDestroyOnLoadManager solves both these issues.